refactor(nofip): move reconciler to separate source file
