+++ cmd/nofip/main.go
@@ -27,71 +27,6 @@ type Config struct {
 	CloudflareToken string
 }
 
-// ReconciliationState defines the state of a resource
-// before the reconciliation process.
-type ReconciliationState struct {
-	// Current is the current state of the resource.
-	Current map[string]bool
-	// Desired is the desired state of the resource.
-	Desired map[string]bool
-}
-
-// ReconciliationPlan defines the plan to reconcile a
-// resource.
-type ReconciliationPlan struct {
-	// Create is the list of records to create.
-	Create map[string]bool
-	// Delete is the list of records to delete.
-	Delete map[string]bool
-}
-
-// Reconciliation is stores the current and desired state
-// of a resource along with a plan to reconcile them.
-type Reconciliation struct {
-	sync.Mutex
-	// State is the state of the resource before the reconciliation.
-	State *ReconciliationState
-	// Plan is the plan to reconcile the resource.
-	Plan *ReconciliationPlan
-}
-
-// ShouldUpdate updates the plan to reconcile the resource.
-func (r *Reconciliation) ShouldUpdate() bool {
-	shouldUpdate := false
-
-	// Check if any new records should be created in the new set.
-	for key := range r.State.Desired {
-		if !r.State.Current[key] {
-			shouldUpdate = true
-			r.Plan.Create[key] = true
-		}
-	}
-
-	// Check if any old records should be removed in the new set.
-	for key := range r.State.Current {
-		if !r.State.Desired[key] {
-			shouldUpdate = true
-			r.Plan.Delete[key] = true
-		}
-	}
-
-	return shouldUpdate
-}
-
-// NewReconciliation creates a new reconciliation object.
-func NewReconciliation() *Reconciliation {
-	return &Reconciliation{
-		State: &ReconciliationState{
-			Current: map[string]bool{},
-			Desired: map[string]bool{},
-		},
-		Plan: &ReconciliationPlan{
-			Create: map[string]bool{},
-			Delete: map[string]bool{},
-		},
-	}
-}
-
 var cfg = &Config{}
 
 var version = "dev"
@@ -147,8 +82,8 @@ geo-DNS, as nofip will only use simple A records.`,
 			return errors.New("must specify at least one edge server")
 		}
 
-		v4Update := NewReconciliation()
-		v6Update := NewReconciliation()
+		v4Update := NewReconciler()
+		v6Update := NewReconciler()
 
 		// Use a waitgroup and a goroutine to resolve the IP addresses of
 		// the edge servers and the current state of the edge record in
@@ -252,7 +187,7 @@ geo-DNS, as nofip will only use simple A records.`,
 		}
 
 		if v6Update.ShouldUpdate() {
-			fmt.Println("[v6] should update")
+			return errors.New("IPv6 is not supported yet")
 			// TODO: Use the cloudflare SDK to create a DNS AAAA record with the IPs.
 		}
